refactor: release line budget in formation-metric owners via declared splits
check:repo-style:changed blocks worsened findings against origin/main, and
every instrumented owner was already at or above the file budget. Each owner
releases budget by extracting one cohesive sub-responsibility (WS server
delivery contracts, group-manager contracts, admin request reading, the
scalar recompute worker registration, RTC status/diagnostics contracts, the
candidate-pair stats reader) with import paths preserved through re-exports,
declared in plans/repo-style-lineages/. The overlay adoption counters move
into overlays-repository.ts so the browser gains no new direct file, and the
group-formation metrics contract moves to packages/shared/rtc so the admin
response fields are fully typed. Zero behavior change; the api-v1 black-box
memory matrix passes 13/13 on this tree.
