+++ apps/api-v1/src/create-rallar-server.ts
@@ -118,6 +118,7 @@ export function createRallarServer(
     now: rtcTopologyOptions.now,
   });
   const adminClientIds = readAdminClientIds();
+
   const topologyManagement = new GroupTopologyManagementService({
     findGroupSnapshotByRef: (ref, cacheOptions) =>
       middleware.groupStateService.readSnapshotAtLeast(ref, cacheOptions ?? {}),
@@ -138,6 +139,7 @@ export function createRallarServer(
   middleware.appGroupInboxService.setTopologyManagementService(
     topologyManagement,
   );
+
   middleware.appGroupInboxService.setRtcRttAppInboxDependencies({
     repository: rttRepository,
     formationMetrics: middleware.groupFormationMetrics.rttMutation,
@@ -192,6 +194,7 @@ export function createRallarServer(
       };
     },
   });
+
   const databaseConfig = readApiV1DatabaseBackendConfig();
   const databasePubSubConfig = readApiV1DatabasePubSubConfig(Deno.env, databaseConfig);
   const emptyWsStatus = {
@@ -205,6 +208,7 @@ export function createRallarServer(
   if (!middleware.appAdminInboxService || !middleware.appCrdtInboxService) {
     throw new Error('Admin database mutations require AppInbox services');
   }
+
   const adminOperations = new AdminOperationsService({
     now,
     serverId: myServerId,
