+++ plans/repo-style-lineages/group-formation-phase0-storm-metrics.json
@@ -1,0 +1,59 @@
+{
+  "version": 1,
+  "lineages": [
+    {
+      "mergeBase": "05cec262b838fb661dcb923152991ce88a28ceac",
+      "source": {
+        "path": "packages/shared/services/WsQueueBoxServerService.ts",
+        "blob": "f510749ffe776db6792df244f832569f387c3ec8"
+      },
+      "targets": ["packages/shared/services/ws-queue-box-server-contracts.ts"]
+    },
+    {
+      "mergeBase": "05cec262b838fb661dcb923152991ce88a28ceac",
+      "source": {
+        "path": "packages/shared/services/WebRtcGroupManager.ts",
+        "blob": "4345aa58768dc318a85972c8bfc0b64f72e2e1a8"
+      },
+      "targets": ["packages/shared/services/webrtc-group-manager-contracts.ts"]
+    },
+    {
+      "mergeBase": "05cec262b838fb661dcb923152991ce88a28ceac",
+      "source": {
+        "path": "packages/shared-server/rallar-system/admin-operations/AdminOperationsService.ts",
+        "blob": "4a90073805e7e4f332d1e4289d35407c05c08dba"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/admin-operations/admin-operations-request-reading.ts"
+      ]
+    },
+    {
+      "mergeBase": "05cec262b838fb661dcb923152991ce88a28ceac",
+      "source": {
+        "path": "apps/api-v1/src/middleware.ts",
+        "blob": "33216b961f45f25681fb8eb167d58b41570b1a5b"
+      },
+      "targets": ["apps/api-v1/src/services/init-api-rtc-topology-scalar-recompute-worker.ts"]
+    },
+    {
+      "mergeBase": "05cec262b838fb661dcb923152991ce88a28ceac",
+      "source": {
+        "path": "packages/shared-web/browser/rallar-rtc-facade.ts",
+        "blob": "6b97e565d50c052984613ad48bcf6b7709e70882"
+      },
+      "targets": [
+        "packages/shared-web/browser/rtc-diagnostics/rallar-rtc-diagnostics-contracts.ts"
+      ]
+    },
+    {
+      "mergeBase": "05cec262b838fb661dcb923152991ce88a28ceac",
+      "source": {
+        "path": "packages/shared-web/browser/rallar-runtime/rtc.ts",
+        "blob": "a189aa3d7024b65a7a0b2aa0855fca89349c1743"
+      },
+      "targets": [
+        "packages/shared-web/browser/rtc-diagnostics/read-selected-candidate-pair-diagnostics.ts"
+      ]
+    }
+  ]
+}
